Extract Avalanche wallet connection functions
